All PASS except GeoDaoTests

+++ mardao-mysql/src/main/java/net/sf/mardao/core/dao/TypeDaoImpl.java
@@ -5,12 +5,10 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Properties;
 import java.util.TreeMap;
 import javax.sql.DataSource;
@@ -26,6 +24,7 @@ import org.springframework.jdbc.BadSqlGrammarException;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.jdbc.support.incrementer.DataFieldMaxValueIncrementer;
 
 public abstract class TypeDaoImpl<T, ID extends Serializable> extends
         DaoImpl<T, ID, Long, Iterable, CoreEntity, CompositeKey> {
@@ -48,6 +47,9 @@ public abstract class TypeDaoImpl<T, ID extends Serializable> extends
     protected NamedParameterJdbcTemplate jdbcTemplate;
     protected SimpleJdbcInsert jdbcInsert;
     private String dialect = DIALECT_DEFAULT;
+    /** Used to generate unique ids */
+    @Autowired
+    protected DataFieldMaxValueIncrementer jdbcIncrementer;
     
     protected TypeDaoImpl(Class<T> type, Class<ID> idType) {
         super(type, idType);
@@ -56,7 +58,9 @@ public abstract class TypeDaoImpl<T, ID extends Serializable> extends
     @Autowired
     public void setDataSource(DataSource dataSource) {
         this.jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
-        this.jdbcInsert = new SimpleJdbcInsert(dataSource).withTableName(getTableName());
+        this.jdbcInsert = new SimpleJdbcInsert(dataSource)
+                .withTableName(getTableName());
+//                .usingGeneratedKeyColumns(getPrimaryKeyColumnName());
         checkTable();
     }
     
@@ -158,10 +162,10 @@ public abstract class TypeDaoImpl<T, ID extends Serializable> extends
         appendColumnDefinition(sql, columnName);
 
         final String className = getColumnClass(columnName).getName();
-        if (Long.class.getName().equals(className)) {
-            sql.append(' ');
-            sql.append(getDataType("AUTO_INCREMENT"));
-        }
+//        if (Long.class.getName().equals(className)) {
+//            sql.append(' ');
+//            sql.append(getDataType("AUTO_INCREMENT"));
+//        }
         sql.append(" PRIMARY KEY");
     }
     
@@ -242,6 +246,9 @@ public abstract class TypeDaoImpl<T, ID extends Serializable> extends
         }
         
         if (Long.class.isAssignableFrom(simpleIdClass)) {
+            if (null == key.getId()) {
+                return null;
+            }
             return (ID) Long.valueOf(key.getId());
         }
         
@@ -336,7 +343,7 @@ public abstract class TypeDaoImpl<T, ID extends Serializable> extends
     
     @Override
     protected final Filter createInFilter(String fieldName, Collection param) {
-        return new Filter(fieldName, "IN", param);
+        return new Filter.IN(fieldName, param);
     }
     
 
@@ -347,6 +354,9 @@ public abstract class TypeDaoImpl<T, ID extends Serializable> extends
     
     @Override
     protected int doDelete(Object parentKey, Iterable<ID> simpleKeys) {
+        if (simpleKeys instanceof ArrayList && ((ArrayList) simpleKeys).isEmpty()) {
+            return 0;
+        }
         final StringBuffer sql = new StringBuffer();
         sql.append("DELETE FROM ");
         sql.append(getTableName());
@@ -423,13 +433,24 @@ public abstract class TypeDaoImpl<T, ID extends Serializable> extends
             }
             sql.append(f.getColumn());
             sql.append(f.getOperation());
-            token = String.format("%s_%d", f.getColumn(), i);
+            token = String.format("%s%d", f.getColumn(), i);
             sql.append(f.getToken(token));
             params.put(token, f.getOperand());
             i++;
         }
         return params;
-    }        
+    }
+    
+    protected void appendLimitOffset(StringBuffer sql, int limit, int offset) {
+        if (0 < limit) {
+            sql.append(" LIMIT ");
+            sql.append(limit);
+            if (0 < offset) {
+                sql.append(" OFFSET ");
+                sql.append(offset);
+            }
+        }
+    }
 
     protected StringBuffer createSelect(boolean keysOnly) {
         final StringBuffer sql = new StringBuffer();
@@ -442,44 +463,42 @@ public abstract class TypeDaoImpl<T, ID extends Serializable> extends
     
     @Override
     protected T doFindByPrimaryKey(Object parentKey, ID simpleKey) {
-        final StringBuffer sql = createSelect(false);
-        final Map<String, Object> params = appendWherePrimaryKeys(sql, (CompositeKey) parentKey, simpleKey);
+        return findUniqueBy(createEqualsFilter(getPrimaryKeyColumnName(), simpleKey));
         
-        LOG.debug("SQL: {} Params: {}", sql.toString(), params);
-        final Map<String, Object> props = jdbcTemplate.queryForMap(sql.toString(), params);
-        
-        final CompositeKey pk = createCoreKey(props);
-        final CoreEntity core = new CoreEntity();
-        core.setPrimaryKey(pk);
-        core.setProperties(props);
-        final T domain = coreToDomain(core);
-        return domain;
+//        final StringBuffer sql = createSelect(false);
+//        final Map<String, Object> params = appendWherePrimaryKeys(sql, (CompositeKey) parentKey, simpleKey);
+//        
+//        LOG.debug("SQL: {} Params: {}", sql.toString(), params);
+//        final Map<String, Object> props = jdbcTemplate.queryForMap(sql.toString(), params);
+//        
+//        final CompositeKey pk = createCoreKey(props);
+//        final CoreEntity core = new CoreEntity();
+//        core.setPrimaryKey(pk);
+//        core.setProperties(props);
+//        final T domain = coreToDomain(core);
+//        return domain;
     }
 
     @Override
     protected Iterable<T> doQueryByPrimaryKeys(Object parentKey, Iterable<ID> simpleKeys) {
-        throw new UnsupportedOperationException("Not supported yet.");
-//        // TODO: get batch with batch size
-//        final Collection<Key> coreKeys = createCoreKeys(parentKey, simpleKeys);
-//        final Map<Key, Entity> entities = getDatastoreService().get(coreKeys);
-//        
-//        final Collection<T> domains = new ArrayList<T>();
-//        T domain;
-//        for (Entity entity : entities.values()) {
-//            domain = coreToDomain(entity);
-//            domains.add(domain);
-//        }
-//        return domains;
+        final ArrayList<ID> values = new ArrayList<ID>();
+        for (ID id : simpleKeys) {
+            values.add(id);
+        }
+        final Filter inFilter = createInFilter(getPrimaryKeyColumnName(), values);
+        return queryIterable(false, 0, -1, null, parentKey, 
+                null, false, null, false, inFilter);
     }
     
     @Override
     protected T findUniqueBy(Filter... filters) {
         StringBuffer sql = createSelect(false);
         Map<String, Object> params = appendWhereFilters(sql, filters);
+        LOG.debug("{} with params {}", sql.toString(), params);
         try {
-        final Map<String, Object> props = jdbcTemplate.queryForMap(sql.toString(), params);
-        final T domain = propsToDomain(props);
-        return domain;
+            final Map<String, Object> props = jdbcTemplate.queryForMap(sql.toString(), params);
+            final T domain = propsToDomain(props);
+            return domain;
         }
         catch (EmptyResultDataAccessException notFound) {
             return null;
@@ -521,28 +540,40 @@ public abstract class TypeDaoImpl<T, ID extends Serializable> extends
 
     @Override
     protected Collection<CompositeKey> persistCore(Iterable<CoreEntity> itrbl) {
-        ArrayList<Map<String, Object>> propsList = new ArrayList<Map<String, Object>>();
+        final Collection<CompositeKey> returnValue = new ArrayList<CompositeKey>();
+        final ArrayList<Map<String, Object>> propsList = new ArrayList<Map<String, Object>>();
+        Long key;
+        CompositeKey parentKey, primaryKey;
+        ID simpleKey;
+        Map<String, Object> props;
         for (CoreEntity core : itrbl) {
-            propsList.add(core.getProperties());
-            LOG.debug("persist props={}", core.getProperties());
-        }
-        final Map[] empty = new Map[propsList.size()];
-        final Map<String, Object>[] batch = propsList.toArray(empty);
-        
-        final int[] simpleKeys = jdbcInsert.executeBatch(batch);
-        
-        final Collection<CompositeKey> returnValue = new ArrayList<CompositeKey>(simpleKeys.length);
-        CompositeKey pk;
-        int i = 0;
-        for (CoreEntity core : itrbl) {
-            pk = core.getPrimaryKey();
-            if (null == pk) {
-                pk = new CompositeKey(null, null, null);
+            props = core.getProperties();
+            simpleKey = (ID) props.get(getPrimaryKeyColumnName());
+            if (null == simpleKey) {
+                Long lid = jdbcIncrementer.nextLongValue();
+                simpleKey = (ID) lid;
+                props.put(getPrimaryKeyColumnName(), simpleKey);
+                LOG.info("generating ID {}", simpleKey);
             }
-            pk.setId((long)simpleKeys[i]);
-            i++;
+            propsList.add(props);
+            LOG.debug("persistCore {}", props);
+            
+            parentKey = null;
+            if (null != getParentKeyColumnName()) {
+                final Long parentId = (Long) props.get(getParentKeyColumnName());
+                parentKey = new CompositeKey(null, parentId, null);
+            }
+            primaryKey = createCoreKey(parentKey, simpleKey);
+            returnValue.add(primaryKey);
         }
         
+        // batch-insert entities with simpleKey assigned
+        if (!propsList.isEmpty()) {
+            final Map[] empty = new Map[propsList.size()];
+            final Map<String, Object>[] batch = propsList.toArray(empty);
+
+            jdbcInsert.executeBatch(batch);
+        }
         return returnValue;
     }
     
@@ -588,6 +619,10 @@ public abstract class TypeDaoImpl<T, ID extends Serializable> extends
         final StringBuffer sql = createSelect(keysOnly);
         
         final Map<String, Object> params = appendWhereFilters(sql, filters);
+        
+        appendLimitOffset(sql, limit, offset);
+        
+        LOG.debug("{} with params {}", sql.toString(), params);
         
         final List<Map<String, Object>> itemsProps = jdbcTemplate.queryForList(sql.toString(), params);
         
@@ -675,6 +710,10 @@ public abstract class TypeDaoImpl<T, ID extends Serializable> extends
                 }
             }
         }
+    }
+
+    public void setJdbcIncrementer(DataFieldMaxValueIncrementer jdbcIncrementer) {
+        this.jdbcIncrementer = jdbcIncrementer;
     }
 
 }
